@@ -23,7 +23,7 @@ import types
 from concurrent.futures._base import Future
 from datetime import timedelta
 from pathlib import Path
-from typing import Any, Literal, Optional, Union
+from typing import Any, Literal
 
 from assertionengine import AssertionOperator
 from overrides import overrides
@@ -798,21 +798,19 @@ class Browser(DynamicCore):
         *_,
         auto_closing_level: AutoClosingLevel = AutoClosingLevel.TEST,
         auto_delete_passed_tracing: bool = False,
-        enable_playwright_debug: Union[
-            PlaywrightLogTypes, bool
-        ] = PlaywrightLogTypes.library,
-        enable_presenter_mode: Union[HighLightElement, bool] = False,
-        external_browser_executable: Optional[dict[SupportedBrowsers, str]] = None,
+        enable_playwright_debug: PlaywrightLogTypes | bool = PlaywrightLogTypes.library,
+        enable_presenter_mode: HighLightElement | bool = False,
+        external_browser_executable: dict[SupportedBrowsers, str] | None = None,
         highlight_on_failure: bool = False,
-        jsextension: Union[list[str], str, None] = None,
-        language: Optional[str] = None,
-        playwright_process_host: Optional[str] = None,
-        playwright_process_port: Optional[int] = None,
-        plugins: Union[list[str], str, None] = None,
+        jsextension: list[str] | str | None = None,
+        language: str | None = None,
+        playwright_process_host: str | None = None,
+        playwright_process_port: int | None = None,
+        plugins: list[str] | str | None = None,
         retry_assertions_for: timedelta = timedelta(seconds=1),
         run_on_failure: str = "Take Screenshot  fail-screenshot-{index}",
-        selector_prefix: Optional[str] = None,
-        show_keyword_call_banner: Optional[bool] = None,
+        selector_prefix: str | None = None,
+        show_keyword_call_banner: bool | None = None,
         strict: bool = True,
         timeout: timedelta = timedelta(seconds=10),
         tracing_group_mode: TracingGroupMode = TracingGroupMode.Full,
@@ -844,7 +842,7 @@ class Browser(DynamicCore):
         self.ROBOT_LIBRARY_LISTENER = self
         self.scope_stack: dict = {}
         self.suite_ids: dict[str, None] = {}
-        self.current_test_id: Optional[str] = None
+        self.current_test_id: str | None = None
         self._playwright_state: PlaywrightState = PlaywrightState(self)
         self._browser_control = Control(self)
         self._assertion_formatter = Formatter(self)
@@ -880,7 +878,7 @@ class Browser(DynamicCore):
             self._playwright_log = None
         else:
             self._playwright_log = self._get_log_file_name()
-        self._playwright: Optional[Playwright] = None
+        self._playwright: Playwright | None = None
         self._auto_closing_level = auto_closing_level
         self.auto_delete_passed_tracing = auto_delete_passed_tracing
         # Parsing needs keywords to be discovered.
@@ -900,14 +898,14 @@ class Browser(DynamicCore):
             self._plugin_keywords = parser.get_plugin_keywords(parsed_plugins)
         else:
             self._plugin_keywords = []
-        self.presenter_mode: Union[HighLightElement, bool] = enable_presenter_mode
+        self.presenter_mode: HighLightElement | bool = enable_presenter_mode
         self.tracing_group_mode = tracing_group_mode
         self._execution_stack: list[dict] = []
         self._running_on_failure_keyword = False
         self.pause_on_failure: set[str] = set()
         self._unresolved_promises: set[Future] = set()
         self._keyword_formatters: dict = {}
-        self._current_loglevel: Optional[str] = None
+        self._current_loglevel: str | None = None
         self.is_test_case_running = False
         self.auto_closing_default_run_before_unload: bool = False
         self.keyword_call_stack: list[KeywordCallStackEntry] = []
@@ -972,9 +970,7 @@ class Browser(DynamicCore):
     def timeout(self):
         return self.scope_stack["timeout"].get()
 
-    def _parse_run_on_failure_keyword(
-        self, keyword: Union[str, None]
-    ) -> DelayedKeyword:
+    def _parse_run_on_failure_keyword(self, keyword: str | None) -> DelayedKeyword:
         if keyword is None or is_falsy(keyword):
             return DelayedKeyword(None, None, (), {})
         parts = keyword.split("  ")
@@ -1007,13 +1003,12 @@ class Browser(DynamicCore):
             response.keywords,
             response.keywordArguments,
             response.keywordDocumentations,
+            strict=False,
         ):
             self._jskeyword_call(component, name, args, doc)
         return component
 
-    def init_js_extension(
-        self, js_extension_path: Union[Path, str]
-    ) -> Response.Keywords:
+    def init_js_extension(self, js_extension_path: Path | str) -> Response.Keywords:
         with self.playwright.grpc_channel() as stub:
             return stub.InitializeExtension(
                 Request().FilePath(
@@ -1191,7 +1186,7 @@ def {name}(self, {", ".join(argument_names_and_default_values_texts)}):
             except ConnectionError as e:
                 logger.trace(f"Browser._start_test connection problem: {e}")
 
-    def _resolve_path(self, attrs: dict) -> Union[Path, None]:
+    def _resolve_path(self, attrs: dict) -> Path | None:
         source = (
             Path(attrs["source"])
             if "source" in attrs and attrs["source"] is not None
@@ -1239,7 +1234,7 @@ def {name}(self, {", ".join(argument_names_and_default_values_texts)}):
         self,
         name: str,
         args: list,
-        source: Union[str, Path, None],
+        source: str | Path | None,
         lineno: int,
         typ: str,
     ) -> KeywordCallStackEntry:
@@ -1551,14 +1546,12 @@ def {name}(self, {", ".join(argument_names_and_default_values_texts)}):
             )
         )
 
-    def get_timeout(self, timeout: Union[timedelta, None]) -> float:
+    def get_timeout(self, timeout: timedelta | None) -> float:
         if timeout is None:
             return self.scope_stack["timeout"].get()
         return self.convert_timeout(timeout)
 
-    def convert_timeout(
-        self, timeout: Union[timedelta, float], to_ms: bool = True
-    ) -> float:
+    def convert_timeout(self, timeout: timedelta | float, to_ms: bool = True) -> float:
         convert = 1000 if to_ms else 1
         if isinstance(timeout, timedelta):
             return timeout.total_seconds() * convert
@@ -1602,7 +1595,7 @@ def {name}(self, {", ".join(argument_names_and_default_values_texts)}):
         return True
 
     @staticmethod
-    def _get_translation(language: Union[str, None]) -> Union[Path, None]:
+    def _get_translation(language: str | None) -> Path | None:
         if not language:
             return None
         discovered_plugins = {
