@@ -12,7 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 import inspect
-from typing import Any, Callable
+from collections.abc import Callable
+from typing import Any
 
 from Browser.utils import logger
 
