@@ -1,7 +1,7 @@
 import { HashRouter, Route, Routes } from 'react-router-dom';
 import DragGame from './draggame';
-import Login from './login';
 import EventsPage from './events';
+import Login from './login';
 import React from 'react';
 
 const App: React.FC = () => {
